refactor(doctor): extract toolsBySender migration helpers
