+++ src/commands/doctor-config-flow.ts
@@ -15,7 +15,6 @@ import { CONFIG_PATH, migrateLegacyConfig, readConfigFileSnapshot } from "../con
 import { collectProviderDangerousNameMatchingScopes } from "../config/dangerous-name-matching.js";
 import { formatConfigIssueLines } from "../config/issue-format.js";
 import { applyPluginAutoEnable } from "../config/plugin-auto-enable.js";
-import { parseToolsBySenderTypedKey } from "../config/types.tools.js";
 import { resolveCommandResolutionFromArgv } from "../infra/exec-command-resolution.js";
 import {
   listInterpreterLikeSafeBins,
@@ -50,15 +49,17 @@ import { listTelegramAccountIds, resolveTelegramAccount } from "../telegram/acco
 import { note } from "../terminal/note.js";
 import { resolveHomeDir } from "../utils.js";
 import {
-  formatConfigPath,
   noteIncludeConfinementWarning,
   noteOpencodeProviderOverrides,
-  resolveConfigPathTarget,
   stripUnknownConfigKeys,
 } from "./doctor-config-analysis.js";
 import { normalizeCompatibilityConfigValues } from "./doctor-legacy-config.js";
 import type { DoctorOptions } from "./doctor-prompter.js";
 import { autoMigrateLegacyStateDir } from "./doctor-state-migrations.js";
+import {
+  maybeRepairLegacyToolsBySenderKeys,
+  scanLegacyToolsBySenderKeys,
+} from "./doctor-tools-by-sender.js";
 
 type TelegramAllowFromUsernameHit = { path: string; entry: string };
 
@@ -1486,120 +1487,6 @@ function maybeRepairExecSafeBinProfiles(cfg: OpenClawConfig): {
     return { config: cfg, changes: [], warnings: [] };
   }
   return { config: next, changes, warnings };
-}
-
-type LegacyToolsBySenderKeyHit = {
-  toolsBySenderPath: Array<string | number>;
-  pathLabel: string;
-  key: string;
-  targetKey: string;
-};
-
-function collectLegacyToolsBySenderKeyHits(
-  value: unknown,
-  pathParts: Array<string | number>,
-  hits: LegacyToolsBySenderKeyHit[],
-) {
-  if (Array.isArray(value)) {
-    for (const [index, entry] of value.entries()) {
-      collectLegacyToolsBySenderKeyHits(entry, [...pathParts, index], hits);
-    }
-    return;
-  }
-  const record = asObjectRecord(value);
-  if (!record) {
-    return;
-  }
-
-  const toolsBySender = asObjectRecord(record.toolsBySender);
-  if (toolsBySender) {
-    const path = [...pathParts, "toolsBySender"];
-    const pathLabel = formatConfigPath(path);
-    for (const rawKey of Object.keys(toolsBySender)) {
-      const trimmed = rawKey.trim();
-      if (!trimmed || trimmed === "*" || parseToolsBySenderTypedKey(trimmed)) {
-        continue;
-      }
-      hits.push({
-        toolsBySenderPath: path,
-        pathLabel,
-        key: rawKey,
-        targetKey: `id:${trimmed}`,
-      });
-    }
-  }
-
-  for (const [key, nested] of Object.entries(record)) {
-    if (key === "toolsBySender") {
-      continue;
-    }
-    collectLegacyToolsBySenderKeyHits(nested, [...pathParts, key], hits);
-  }
-}
-
-function scanLegacyToolsBySenderKeys(cfg: OpenClawConfig): LegacyToolsBySenderKeyHit[] {
-  const hits: LegacyToolsBySenderKeyHit[] = [];
-  collectLegacyToolsBySenderKeyHits(cfg, [], hits);
-  return hits;
-}
-
-function maybeRepairLegacyToolsBySenderKeys(cfg: OpenClawConfig): {
-  config: OpenClawConfig;
-  changes: string[];
-} {
-  const next = structuredClone(cfg);
-  const hits = scanLegacyToolsBySenderKeys(next);
-  if (hits.length === 0) {
-    return { config: cfg, changes: [] };
-  }
-
-  const summary = new Map<string, { migrated: number; dropped: number; examples: string[] }>();
-  let changed = false;
-
-  for (const hit of hits) {
-    const toolsBySender = asObjectRecord(resolveConfigPathTarget(next, hit.toolsBySenderPath));
-    if (!toolsBySender || !(hit.key in toolsBySender)) {
-      continue;
-    }
-    const row = summary.get(hit.pathLabel) ?? { migrated: 0, dropped: 0, examples: [] };
-
-    if (toolsBySender[hit.targetKey] === undefined) {
-      toolsBySender[hit.targetKey] = toolsBySender[hit.key];
-      row.migrated++;
-      if (row.examples.length < 3) {
-        row.examples.push(`${hit.key} -> ${hit.targetKey}`);
-      }
-    } else {
-      row.dropped++;
-      if (row.examples.length < 3) {
-        row.examples.push(`${hit.key} (kept existing ${hit.targetKey})`);
-      }
-    }
-    delete toolsBySender[hit.key];
-    summary.set(hit.pathLabel, row);
-    changed = true;
-  }
-
-  if (!changed) {
-    return { config: cfg, changes: [] };
-  }
-
-  const changes: string[] = [];
-  for (const [pathLabel, row] of summary) {
-    if (row.migrated > 0) {
-      const suffix = row.examples.length > 0 ? ` (${row.examples.join(", ")})` : "";
-      changes.push(
-        `- ${pathLabel}: migrated ${row.migrated} legacy key${row.migrated === 1 ? "" : "s"} to typed id: entries${suffix}.`,
-      );
-    }
-    if (row.dropped > 0) {
-      changes.push(
-        `- ${pathLabel}: removed ${row.dropped} legacy key${row.dropped === 1 ? "" : "s"} where typed id: entries already existed.`,
-      );
-    }
-  }
-
-  return { config: next, changes };
 }
 
 async function maybeMigrateLegacyConfig(): Promise<string[]> {
